Fix type mismatch in windows_console.rs and add module documentation
- Fixed: Type mismatch error where stdout as isize was compared to *mut c_void INVALID_HANDLE_VALUE
- Changed: Direct pointer comparison stdout == INVALID_HANDLE_VALUE
- Added: Module documentation comments for workflows and knowledge modules
- Tested: MCP compliance tests pass with 48 tools
- Built: Windows x86_64 exe successfully compiled

+++ src/bridge/windows_console.rs
@@ -17,7 +17,7 @@ pub fn attach_console() {
         let stdout: *mut std::ffi::c_void = GetStdHandle(STD_OUTPUT_HANDLE);
         
         // If stdout is invalid or null, we need to attach to parent console
-        if stdout.is_null() || stdout as isize == INVALID_HANDLE_VALUE {
+        if stdout.is_null() || stdout == INVALID_HANDLE_VALUE {
             // Try to attach to parent console
             let result = AttachConsole(ATTACH_PARENT_PROCESS);
